desktop/requirements.txt
PyQt5>=5.15
PyOpenGL>=3.1.6
pyserial>=3.5
numpy>=1.24
scipy>=1.10
ezdxf>=1.0
trimesh>=4.0
